V2 of App and Photon Code
V2 changes the way in which measurements are taken.  Instead of taking a measurement every one second, the firmware now only measures when a measurement is called for.  The App is changed to call a new cloud function to perform a measurement, vs the old cloud variable to read out the result.